Extract DropletInfo and DropletList classes from API::AllDroplets
